Accepted: Destroying Asteroids

+++ 2126-destroying-asteroids/2126-destroying-asteroids.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean asteroidsDestroyed(int k, int[] nums) {
         Arrays.sort(nums);
-        int tar = k;
+        long tar = k;
 
         for(int num : nums) {
             if(num <= tar) {
